requirements.txt
alembic==1.0.7
Click==7.0
dominate==2.3.5
eventlet==0.24.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.3.1
Flask-SocketIO==3.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
python-engineio==3.3.0
python-socketio==3.1.1
six==1.12.0
SQLAlchemy==1.2.17
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1


#Heroku dependencies
psycopg2==2.7.3.1
gunicorn==19.9.0